refactor: 도메인 -> View 의존성 제거
