+++ src/main/java/racingcar/constant/CarConstant.java
@@ -6,5 +6,6 @@ public class CarConstant {
     public static final int MAX_NAME_LENGTH = 5;
     public static final int MOVE_MINIMUM_VALUE = 4;
     public static final int CAR_DISTANCE_INIT_VALUE = 0;
-    public static final String winnerNameDelimiter = ", ";
+    public static final String WINNER_NAME_DELIMITER = ", ";
+    public static final String COMMA = ",";
 }
