+++ src/main/java/racingcar/domain/Car.java
@@ -3,7 +3,7 @@ package racingcar.domain;
 import racingcar.constant.CarConstant;
 
 public class Car {
-    private String name;
+    private final String name;
     private int distance;
 
     public Car(String name, int distance) {
