+++ src/test/java/racingcar/domain/CarTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racingcar.constant.Message;
+import racingcar.view.Message;
 
 import static org.junit.jupiter.api.Assertions.*;
